Added 4.0.4_r2.1 build
Since building from 'master' was giving me R_ARM_REL32 linking issues
with the PDK gcc compiler, I added a set of build scripts for the last
build (4.0) before the linker bug showed up that we are seeing in
master. I expect other compilers to be able to work around this issue
for webOS 3.0.5, and other platforms to not have this issue at all.