Initial work on supporting Zygote on webOS
The Zygote stub is down to only requiring a single enum value for ADDR_NO_RANDOMIZE from personality.h
We could probably fix this by force including a header with the a zero value for ADDR_NO_RANDOMIZE for the Zygote.cpp file.

The rest of the additional changes are an attempt to build software required to build android_runtime, which app_process links against.

We are going to have to decide how much of android_runtime we require, and how much we need to just rewrite from scratch to support webOS.
Probably need a webos_runtime library to either supplement, or replace android_runtime. For now, I'm trying to build android_runtime, which depends on
other libraries like skia, webp, jpeg, ect.

Much of the new additions won't even compile, so they are wrapped in a WORK_IN_PROGRESS conditial statement for cmake.